Added navigation drawer
*NOTE*
The titles are not currently working. I believe this is due to the onAttach not being implemented in the respective fragments.

Also, there is lag in the navigation drawer. The only fix i've found to this is to delay the starting of the activity, which is very hacky and not optimal.